Refactor IstioWizard logic - Extract Istio config built in a separate class
